refactor of the  GET_ROAD_METADATA function to add new params as a way of removing cyclic dependency
